Make dedicated StatelessWidget for displaying contact detail
